Move back to a GNU attribute
I would *love* to use a C++11 attribute, but they seem to be broken when they
apply to types. It's not just the prohibition whose removal I reverted here;
the parser seems not to be able to apply them to components of a type (and has
no facility for doing so in general).